+++ changelog.md
@@ -15,6 +15,7 @@
 - ✨ Add parameter IO support for more formats supported by pandas (#896)
 - 👌 Apply IRF shift in coherent artifact megacomplex (#992)
 - 👌 Added IRF shift to result dataset (#994)
+- 👌 Improve Result, Parameter and ParameterGroup markdown (#1012)
 
 ### 🩹 Bug fixes